GOCBC-1830: Update student record example project structure
