+++ modules/devguide/examples/go/student-record/internal/logger.go
@@ -1,4 +1,4 @@
-package main
+package internal
 
 import (
 	"os"
